#30 Add client side part to admin bar component
