+++ src/components/admin/admin-bar/index.ts
@@ -1,0 +1,1 @@
+export { AdminBarServer as AdminBar } from './admin-bar.server';
